+++ plugins/provisioners/ansible/cap/guest/arch/ansible_install.rb
@@ -1,3 +1,4 @@
+require_relative "../../../errors"
 
 module VagrantPlugins
   module Ansible
@@ -6,9 +7,13 @@ module VagrantPlugins
         module Arch
           module AnsibleInstall
 
-            def self.ansible_install(machine)
-              machine.communicate.sudo("pacman -Syy --noconfirm")
-              machine.communicate.sudo("pacman -S --noconfirm ansible")
+            def self.ansible_install(machine, install_mode, ansible_version)
+              if install_mode == :pip
+                raise Ansible::Errors::AnsiblePipInstallIsNotSupported
+              else
+                machine.communicate.sudo("pacman -Syy --noconfirm")
+                machine.communicate.sudo("pacman -S --noconfirm ansible")
+              end
             end
 
           end