+++ plugins/provisioners/ansible/cap/guest/fedora/ansible_install.rb
@@ -1,3 +1,5 @@
+require_relative "../facts"
+require_relative "../pip/pip"
 
 module VagrantPlugins
   module Ansible
@@ -6,16 +8,20 @@ module VagrantPlugins
         module Fedora
           module AnsibleInstall
 
-            def self.ansible_install(machine)
-              if dnf?(machine)
-                machine.communicate.sudo("dnf -y install ansible")
+            def self.ansible_install(machine, install_mode, ansible_version)
+              if install_mode == :pip
+                pip_setup machine
+                Pip::pip_install machine, "ansible", ansible_version
               else
-                machine.communicate.sudo("yum -y install ansible")
+                machine.communicate.sudo "#{Facts::rpm_package_manager} -y install ansible"
               end
             end
 
-            def self.dnf?(machine)
-              machine.communicate.test("/usr/bin/which -s dnf")
+            private
+
+            def self.pip_setup(machine)
+              machine.communicate.sudo "#{Facts::rpm_package_manager(machine)} install -y curl gcc gmp-devel libffi-devel openssl-devel python-crypto python-devel python-dnf python-setuptools redhat-rpm-config"
+              Pip::get_pip machine
             end
 
           end