+++ plugins/provisioners/ansible/cap/guest/redhat/ansible_install.rb
@@ -1,3 +1,5 @@
+require_relative "../facts"
+require_relative "../pip/pip"
 
 module VagrantPlugins
   module Ansible
@@ -6,17 +8,29 @@ module VagrantPlugins
         module RedHat
           module AnsibleInstall
 
-            def self.ansible_install(machine)
-              epel = machine.communicate.execute("#{yum_dnf(machine)} repolist epel | grep -q epel", :error_check => false)
-              if epel != 0
-                machine.communicate.sudo('sudo rpm -i https://dl.fedoraproject.org/pub/epel/epel-release-latest-`rpm -E %dist | sed -n \'s/.*el\([0-9]\).*/\1/p\'`.noarch.rpm')
+            def self.ansible_install(machine, install_mode, ansible_version)
+              if install_mode == :pip
+                pip_setup machine
+                Pip::pip_install machine, "ansible", ansible_version
+              else
+                ansible_rpm_install machine
               end
-
-              machine.communicate.sudo("#{yum_dnf(machine)} -y --enablerepo=epel install ansible")
             end
 
-            def self.yum_dnf(machine)
-              machine.communicate.test("/usr/bin/which -s dnf") ? "dnf" : "yum"
+            private
+
+            def self.ansible_rpm_install(machine)
+              epel = machine.communicate.execute "#{yum_dnf(machine)} repolist epel | grep -q epel", error_check: false
+              if epel != 0
+                machine.communicate.sudo 'sudo rpm -i https://dl.fedoraproject.org/pub/epel/epel-release-latest-`rpm -E %dist | sed -n \'s/.*el\([0-9]\).*/\1/p\'`.noarch.rpm'
+              end
+
+              machine.communicate.sudo "#{Facts::rpm_package_manager} -y --enablerepo=epel install ansible"
+            end
+
+            def self.pip_setup(machine)
+              machine.communicate.sudo("#{Facts::rpm_package_manager(machine)} install -y curl gcc libffi-devel openssl-devel python-crypto python-devel python-setuptools")
+              Pip::get_pip machine
             end
 
           end