+++ plugins/provisioners/ansible/cap/guest/suse/ansible_install.rb
@@ -6,8 +6,12 @@ module VagrantPlugins
         module SUSE
           module AnsibleInstall
 
-            def self.ansible_install(machine)
-              machine.communicate.sudo("zypper --non-interactive --quiet install ansible")
+            def self.ansible_install(machine, install_mode, ansible_version)
+              if install_mode == :pip
+                raise Ansible::Errors::AnsiblePipInstallIsNotSupported
+              else
+                machine.communicate.sudo("zypper --non-interactive --quiet install ansible")
+              end
             end
 
           end