+++ plugins/provisioners/ansible/config/guest.rb
@@ -9,12 +9,14 @@ module VagrantPlugins
         attr_accessor :provisioning_path
         attr_accessor :tmp_path
         attr_accessor :install
+        attr_accessor :install_mode
         attr_accessor :version
 
         def initialize
           super
 
           @install           = UNSET_VALUE
+          @install_mode      = UNSET_VALUE
           @provisioning_path = UNSET_VALUE
           @tmp_path          = UNSET_VALUE
           @version           = UNSET_VALUE
@@ -24,6 +26,7 @@ module VagrantPlugins
           super
 
           @install           = true                   if @install           == UNSET_VALUE
+          @install_mode      = :default               if @install_mode      == UNSET_VALUE
           @provisioning_path = "/vagrant"             if provisioning_path  == UNSET_VALUE
           @tmp_path          = "/tmp/vagrant-ansible" if tmp_path           == UNSET_VALUE
           @version           = ""                     if @version           == UNSET_VALUE
@@ -31,6 +34,12 @@ module VagrantPlugins
 
         def validate(machine)
           super
+
+          if @install_mode.to_s.to_sym == :pip
+            @install_mode = :pip
+          else
+            @install_mode = :default
+          end
 
           { "ansible local provisioner" => @errors }
         end