+++ CHANGELOG.md
@@ -45,6 +45,8 @@ IMPROVEMENTS:
   - commands/login: Print a warning with both the environment variable and
       local login token are present [GH-7206, GH-7219]
   - communicators/winrm: Upgrade to latest WinRM gems [GH-6922]
+  - provisioners/ansible_local: Allow to install Ansible from pip,
+      with version selection capability [GH-6654, GH-7167]
   - provisioners/ansible_local: Use `provisioning_path` as working directory
       for `ansible-galaxy` execution
   - provisioners/ansible(both provisioners): Add basic config